feat: move helpers function to utils and create object with languages
