+++ scripts/components/Main.js
@@ -1,4 +1,4 @@
-import createElement from './createElement.js';
+import createElement from '../utils/createElement.js';
 
 export default class Main {
   constructor() {
