+++ scripts/languages/languages.js
@@ -1,0 +1,7 @@
+import ru from './ru.js';
+import eng from './eng.js';
+
+export default {
+    'ru': ru,
+    'eng': eng
+}
